ows/requirements.txt
uvicorn
numpy
rasterio
pyyaml
tqdm
pyproj
psycopg2-binary
shapely
geoalchemy2
sqlalchemy